Create function redirect to booking view
