+++ FlightReservations/App.tsx
@@ -1,15 +1,16 @@
-import {View} from 'react-native';
 import React from 'react';
-import Login from './src/screens/Login/Login';
+import 'react-native-gesture-handler';
+import {NavigationContainer} from '@react-navigation/native';
+import Navigation from './src/components/Navigation/Navigation';
 import {store} from './src/store/store';
 import {Provider} from 'react-redux';
 
 export default function App() {
   return (
     <Provider store={store}>
-      <View>
-        <Login />
-      </View>
+      <NavigationContainer>
+        <Navigation />
+      </NavigationContainer>
     </Provider>
   );
 }
