+++ FlightReservations/src/components/Navigation/Navigation.tsx
@@ -1,25 +1,28 @@
 import React from 'react';
 import {createStackNavigator} from '@react-navigation/stack';
-import {Login} from '../../screens/Login/Login';
-import {MyFlight} from '../../screens/MyFlight/MyFlight';
-import {Booking} from '../../screens/Booking/Booking';
-import {Register} from '../../screens/Register/Register';
+import Login from '../../screens/Login/Login';
+import MyFlight from '../../screens/MyFlight/MyFlight';
+import Booking from '../../screens/Booking/Booking';
+import Register from '../../screens/Register/Register';
 
 const Stack = createStackNavigator();
 
 function Navigation() {
   return (
     <Stack.Navigator>
-      {/* <Stack.Screen name="Register" component={Register} />
-      <Stack.Screen name="Login" component={Login} options={{title: ''}} /> */}
-      <Stack.Screen name="MyFlight" component={MyFlight} />
       <Stack.Screen
-        name="Booking"
-        component={Booking}
+        name="Register"
+        component={Register}
         options={{headerShown: false}}
       />
+      <Stack.Screen name="Login" component={Login} options={{title: ''}} />
+      <Stack.Screen
+        name="MyFlight"
+        component={MyFlight}
+        options={{title: ''}}
+      />
+      <Stack.Screen name="Booking" component={Booking} options={{title: ''}} />
     </Stack.Navigator>
   );
 }
-
 export default Navigation;
