+++ FlightReservations/src/screens/Booking/Booking.tsx
@@ -1,10 +1,12 @@
 import React from 'react';
 import {View, Text} from 'react-native';
 
-export const Booking = () => {
+const Booking = () => {
   return (
     <View>
       <Text>Booking</Text>
     </View>
   );
 };
+
+export default Booking;
