+++ FlightReservations/src/screens/MyFlight/styleMyflights.tsx
@@ -1,0 +1,11 @@
+import {StyleSheet} from 'react-native';
+export const styles = StyleSheet.create({
+  bntFloating: {
+    backgroundColor: '#5974F5',
+    position: 'absolute',
+    borderRadius: 50,
+    padding: 10,
+    bottom: 15,
+    left: '40%',
+  },
+});
